♻️ Refactor `store`
